refactor: remove SUIT font and login main refactoring
