+++ frontend-hybrid/yakal/lib/main.dart
@@ -3,15 +3,15 @@ import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:flutter_native_splash/flutter_native_splash.dart';
 import 'package:get/get.dart';
 import 'package:kakao_flutter_sdk_user/kakao_flutter_sdk_user.dart';
-import 'package:yakal/screens/Auth/auth_finish_screen.dart';
-import 'package:yakal/screens/Auth/enter_nickname_screen.dart';
-import 'package:yakal/screens/Auth/identification_entry_screen.dart';
-import 'package:yakal/screens/Auth/identification_result_screen.dart';
-import 'package:yakal/screens/Auth/identification_screen.dart';
-import 'package:yakal/screens/Auth/login_entry_screen.dart';
-import 'package:yakal/screens/Auth/login_terms_screen.dart';
-import 'package:yakal/screens/Auth/select_mode_screen.dart';
 import 'package:yakal/screens/Home/home_screen.dart';
+import 'package:yakal/screens/Login/LoginEntry/screen.dart';
+import 'package:yakal/screens/Login/auth_finish_screen.dart';
+import 'package:yakal/screens/Login/enter_nickname_screen.dart';
+import 'package:yakal/screens/Login/identification_entry_screen.dart';
+import 'package:yakal/screens/Login/identification_result_screen.dart';
+import 'package:yakal/screens/Login/identification_screen.dart';
+import 'package:yakal/screens/Login/login_terms_screen.dart';
+import 'package:yakal/screens/Login/select_mode_screen.dart';
 import 'package:yakal/screens/Profile/profile_screen.dart';
 import 'package:yakal/screens/Setting/setting_screen.dart';
 import 'package:yakal/screens/Setting/setting_signout_screen.dart';
