+++ frontend-hybrid/yakal/lib/screens/Login/LoginEntry/style.dart
@@ -1,0 +1,26 @@
+import 'package:flutter/material.dart';
+
+import '../../../utilities/style/color_styles.dart';
+
+class LoginEntryStyle {
+  static const blackSubtitle = TextStyle(
+    color: ColorStyles.black,
+    fontSize: 20,
+    fontWeight: FontWeight.w500,
+    height: 1.685,
+  );
+
+  static const blueBoldSubtitle = TextStyle(
+    color: ColorStyles.main,
+    fontSize: 20,
+    fontWeight: FontWeight.w700,
+    height: 1.685,
+  );
+
+  static const title = TextStyle(
+    color: ColorStyles.main,
+    fontSize: 40,
+    fontWeight: FontWeight.w700,
+    height: 1.6,
+  );
+}
