+++ frontend-hybrid/yakal/lib/screens/Login/identification_result_screen.dart
@@ -5,8 +5,8 @@ import 'package:yakal/utilities/api/api.dart';
 
 import '../../utilities/enum/login_process.dart';
 import '../../utilities/style/color_styles.dart';
-import '../../widgets/Auth/back_confirm_dialog.dart';
-import '../../widgets/Auth/login_progress_bar.dart';
+import '../../widgets/Login/back_confirm_dialog.dart';
+import '../../widgets/Login/login_progress_bar.dart';
 
 class IdentificationResultScreen extends StatelessWidget {
   const IdentificationResultScreen({super.key});
