+++ frontend-hybrid/yakal/lib/screens/Login/login_terms_screen.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:get/get.dart';
-import 'package:yakal/screens/Auth/terms_detail_screen.dart';
+import 'package:yakal/screens/Login/terms_detail_screen.dart';
 import 'package:yakal/utilities/enum/login_process.dart';
-import 'package:yakal/widgets/Auth/auth_check_button.dart';
-import 'package:yakal/widgets/Auth/back_confirm_dialog.dart';
-import 'package:yakal/widgets/Auth/login_progress_bar.dart';
+import 'package:yakal/widgets/Login/auth_check_button.dart';
+import 'package:yakal/widgets/Login/back_confirm_dialog.dart';
+import 'package:yakal/widgets/Login/login_progress_bar.dart';
 
 import '../../utilities/style/color_styles.dart';
