+++ frontend-hybrid/yakal/lib/screens/Login/terms_detail_screen.dart
