+++ frontend-hybrid/yakal/lib/widgets/Login/auth_check_button.dart
