+++ frontend-hybrid/yakal/lib/widgets/Login/login_progress_bar.dart
