+++ frontend-hybrid/yakal/pubspec.yaml
@@ -63,23 +63,3 @@ flutter:
           weight: 800
         - asset: assets/fonts/Pretendard-Black.otf
           weight: 900
-    - family: SUIT
-      fonts:
-        - asset: assets/fonts/SUIT-Thin.otf
-          weight: 100
-        - asset: assets/fonts/SUIT-ExtraLight.otf
-          weight: 200
-        - asset: assets/fonts/SUIT-Light.otf
-          weight: 300
-        - asset: assets/fonts/SUIT-Regular.otf
-          weight: 400
-        - asset: assets/fonts/SUIT-Medium.otf
-          weight: 500
-        - asset: assets/fonts/SUIT-SemiBold.otf
-          weight: 600
-        - asset: assets/fonts/SUIT-Bold.otf
-          weight: 700
-        - asset: assets/fonts/SUIT-ExtraBold.otf
-          weight: 800
-        - asset: assets/fonts/SUIT-Heavy.otf
-          weight: 900
